Split dedup and stats into lib and cli for 1.0
